refactor to used es6 syntax
