+++ index.html
@@ -45,6 +45,6 @@
           <p>&copy;awesomebooks || @2022</p>
          
         </footer>
-          <script type="module" src="index.js"></script>
+          <script type="module" src="/js/index.js"></script>
 </body>
 </html>
